增加目录导入功能

+++ src/page/resource/resource-manage/resource-item/resourceCatalog.vue
@@ -2,46 +2,68 @@
   <div>
     <div class="seach_condition">
       <Input v-model="searchName" placeholder="输入搜索名称" style="width: 200px"></Input>
-      <div class="search_button">
-        <i-button class="marginLeft">导入</i-button>
-      </div>
     </div>
     <tree-table ref="treeTable" :items='data3' :columns='dataColumns' @on-row-click="rowClick" @on-expand-click="loadData"></tree-table>
     <Modal v-model="editItemModal" :closable='false' :mask-closable=false :width="500">
       <h3 slot="header" style="color:#2D8CF0">修改目录</h3>
-      <Form ref="editItemForm" :model="editItemForm" :label-width="100" label-position="right" :rules="passwordValidate">
-        <FormItem label="id" prop="oldPass" :error="oldPassError">
-          <Input v-model="editItemForm.id" placeholder="请输入现在使用的密码"></Input>
+      <Form ref="editItemForm" :model="editItemForm" :label-width="100" label-position="right">
+        <FormItem label="id" prop="id">
+          <Input v-model="editItemForm.id" placeholder="数据id"></Input>
         </FormItem>
-        <FormItem label="数据名称" prop="newPass">
-          <Input v-model="editItemForm.name" placeholder="请输入新密码，至少6位字符"></Input>
+        <FormItem label="数据名称" prop="name">
+          <Input v-model="editItemForm.name" placeholder="请输入输入数据名称"></Input>
         </FormItem>
         <FormItem label="数据描述" prop="description">
-          <Input v-model="editItemForm.description" placeholder="请再次输入新密码"></Input>
+          <Input v-model="editItemForm.description" placeholder="请输入数据描述"></Input>
         </FormItem>
         <FormItem label="数据类型" prop="type">
           <Select v-model="editItemForm.type">
-            <Option value="beijing">New York</Option>
-            <Option value="shanghai">London</Option>
-            <Option value="shenzhen">Sydney</Option>
+            <Option value="1">目录</Option>
+            <Option value="2">矢量</Option>
+            <Option value="3">文本</Option>
+            <Option value="4">720</Option>
           </Select>
         </FormItem>
         <FormItem label="数据更新时间" prop="time">
-          <Input v-model="editItemForm.time" placeholder="请再次输入新密码"></Input>
+          <Input v-model="editItemForm.time" placeholder="数据更新时间"></Input>
         </FormItem>
       </Form>
       <div slot="footer">
-        <Button type="text" @click="cancelEditPass">取消</Button>
-        <Button type="primary" :loading="savePassLoading" @click="saveEditPass">保存</Button>
+        <Button type="text" @click="cancelModal">取消</Button>
+        <Button type="primary" @click="saveEditPass">保存</Button>
       </div>
+    </Modal>
+    <Modal v-model="importModal" :closable='false' :mask-closable="false" :width="500" @on-ok="saveImport" @on-cancel="cancelModal">
+      <h3 slot="header" style="color:#2D8CF0">目录导入</h3>
+      <Form label-position="left" :label-width="80">
+        <FormItem label="导入类型">
+          <Select v-model="importFile.type" placeholder="请选择导入类型">
+            <Option value="1">增量导入</Option>
+            <Option value="2">全量导入</Option>
+          </Select>
+        </FormItem>
+        <FormItem label="选择文件">
+          <Upload :action="`${uploadUrl}/sys/file/upload.do`" with-credentials :on-success="handleSuccessUpload">
+            <Button type="ghost" icon="ios-cloud-upload-outline">请选择</Button>
+          </Upload>
+        </FormItem>
+        <div class="importSlot">
+          <div class="importSlotTitle">导入须知</div>
+          <p>1、导入文件大小不超过2MB.</p>
+          <p>2、支持Microsoft Office Excel的xls和xlsx文件,模板
+            <a href="/mouldFile/ms_branch_struct.xlsx" download="ms_branch_struct.xlsx">点此下载.</a>
+          </p>
+        </div>
+      </Form>
     </Modal>
   </div>
 </template>
 
 <script>
-import TreeTable from '@/components/tree-table/index'
+import { url } from '@/api/config.js'
+import TreeTable from '@/components/my-tree/index'
 import { getDateTree } from '@/api/system'
-import { getAreaList, getMsTabDatainfoById, updateMsTabDatainfo, deleteMsTabDatainfo, importMsTabFile } from '@/api/catalog'
+import { getAreaList, getAreaCatalog, getMsTabDatainfoById, updateMsTabDatainfo, deleteMsTabDatainfo, importMsTabFile } from '@/api/catalog'
 
 export default {
   components: {
@@ -53,6 +75,13 @@ export default {
       code: '',
       searchName: '',
       editItemModal: false,
+      importModal: false,
+      uploadUrl: url,
+      importFile: {
+        file: '',
+        type: '',
+        areacode: 500000
+      },
       editItemForm: {
         id: '',
         name: '',
@@ -69,21 +98,25 @@ export default {
           title: '数据名称',
           key: 'title'
         }, {
-          title: '数据类型',
+          title: '数据编码',
           key: 'id',
           sortable: true
         }, {
           title: '排序',
-          key: 'sort'
+          key: 'listorder'
         }, {
           title: '更新时间',
-          key: 'time'
+          key: 'updatetime'
         }, {
           title: '操作',
           type: 'action',
           actions: [{
             type: 'primary',
             text: '编辑'
+          }, {
+            type: 'success',
+            text: '导入',
+            key: 'parentId'
           }, {
             type: 'error',
             text: '删除'
@@ -93,20 +126,43 @@ export default {
     }
   },
   created() {
-    this._getAreaList()
+    this._getAreaCatalog()
   },
   methods: {
-    cancelEditPass() {
+    cancelModal() {
       this.editItemModal = false
+      this.importModal = false
     },
     saveEditPass() {
       this._updateMsTabDatainfo()
     },
+    handleSuccessUpload(response, file) {
+      this.importFile.file = response.data
+    },
+    // 上传文件
+    saveImport() {
+      if (this.importFile.type === '') {
+        this._mm.errorTips('请选择导入类型')
+      } else if (this.importFile.file === '') {
+        this._mm.errorTips('请选择上传文件')
+      } else {
+        this._importMsTabFile(this.importFile)
+      }
+    },
     rowClick(item, e) {
       if (e.target.innerText === '删除') {
-        this._deleteMsTabDatainfo(item.id)
+        this.$Modal.confirm({
+          title: '提示',
+          content: '级联删除且删除的记录不可恢复，继续吗？',
+          onOk: () => {
+            this._deleteMsTabDatainfo(item.id)
+          }
+        })
       } else if (e.target.innerText === '编辑') {
         this._getMsTabDatainfoById(item.id)
+      } else if (e.target.innerText === '导入') {
+        this.importFile.areacode = item.id
+        this.importModal = true
       }
     },
     loadData(item, index) {
@@ -126,6 +182,12 @@ export default {
         } else {
           this._mm.errorTips(res.message)
         }
+      })
+    },
+    _getAreaCatalog() {
+      getAreaCatalog().then(res => {
+        res.map(v => v._checked = false)
+        this.data3 = res
       })
     },
     _getDateTree(id) {
@@ -150,6 +212,13 @@ export default {
     _getMsTabDatainfoById(id) {
       getMsTabDatainfoById(id).then(res => {
         if (res.code === 20000) {
+          this.editItemForm = {
+            id: res.data.id,
+            name: res.data.name,
+            type: res.data.type,
+            description: res.data.description,
+            time: this._mm.formatDate(res.data.updatetime)
+          }
           this.editItemModal = true
         } else {
           this._mm.errorTips(`获取${res.message}`)
@@ -162,6 +231,16 @@ export default {
           this._mm.successTips(`修改${res.message}`)
         } else {
           this._mm.errorTips(`修改${res.message}`)
+        }
+      })
+    },
+    _importMsTabFile(data) {
+      importMsTabFile(data).then(res => {
+        if (res.code === 20000) {
+          this._mm.successTips(res.message)
+          this._getAreaCatalog()
+        } else {
+          this._mm.errorTips(res.message)
         }
       })
     }
